try to setup upgrade cron job

+++ packer/ubuntu/22.04/minecraft-local-storage/files/upgrade_minecraft.sh
@@ -1,34 +1,45 @@
 now=`date +"%Y-%m-%d"`
-  
+
 DOWNLOAD_URL=$(curl -H "Accept-Encoding: identity" -H "Accept-Language: en" -s -L -A "Mozilla/4.0 (compatible; MSIE 6.0; Windows NT 5.1; BEDROCK-UPDATER)" https://minecraft.net/en-us/download/server/bedrock/ |  grep -o 'https://minecraft.azureedge.net/bin-linux/[^"]*')
-sudo wget $DOWNLOAD_URL -O bedrock-server.zip
 
-# stop minecraft server
-sudo systemctl stop mcbedrock
+CURRENT_VERSION=$(echo "$DOWNLOAD_URL" | grep -oP 'bin-linux/\K[^/]+')
+VERSION_FILE="current_version.txt"
+TEMP_DIR=$(mktemp -d)
+SERVER_DIR=/home/mcserver/minecraft_bedrock
 
-# backup entire minecraft directory
-tar -cvpzf minecraft-full-backup${now}.tar.gz /home/mcserver/minecraft_bedrock
-sudo cp minecraft-full-backup${now}.tar.gz /mnt/homenas/minecraft/backups/
+if [ -f "$VERSION_FILE" ]; then
+  LAST_VERSION=$(cat "$VERSION_FILE")
+else
+  LAST_VERSION=""
+fi
 
-# copy config files to temporary backup folder
-sudo mkdir /home/mcserver/minecraft_backup${now}
-sudo mkdir /home/mcserver/minecraft_backup${now}/worlds
-# copy server config file
-sudo cp /home/mcserver/minecraft_bedrock/server.properties /home/mcserver/minecraft_backup${now}/
-# copy worlds
-sudo cp -r /home/mcserver/minecraft_bedrock/worlds /home/mcserver/minecraft_backup${now}/worlds/
-# copy allow list
-sudo cp /home/mcserver/minecraft_bedrock/allowlist.json /home/mcserver/minecraft_backup${now}/
+if [ "$CURRENT_VERSION" != "$LAST_VERSION" ]; then
+  
+  echo "New version available: $CURRENT_VERSION. Downloading and installing..."
 
-sudo unzip -o ./bedrock-server.zip -d /home/mcserver/minecraft_bedrock/
+  sudo wget $DOWNLOAD_URL -O bedrock-server.zip
 
-# restore server config
-sudo cp /home/mcserver/minecraft_backup${now}/server.properties /home/mcserver/minecraft_bedrock/
-# restore worlds
-sudo cp -r /home/mcserver/minecraft_backup${now}/worlds /home/mcserver/minecraft_bedrock/worlds/
-# restore allow list
-sudo cp /home/mcserver/minecraft_backup${now}/allowlist.json /home/mcserver/minecraft_bedrock/
+  # stop minecraft server
+  sudo systemctl stop mcbedrock
 
-sudo chown -R mcserver: /home/mcserver
+  sudo cp "$SERVER_DIR/server.properties" "$TEMP_DIR"
+  sudo cp "$SERVER_DIR/allowlist.json" "$TEMP_DIR"
+  sudo cp "$SERVER_DIR/permissions.json" "$TEMP_DIR"
 
-sudo systemctl start mcbedrock
+  sudo unzip -oq ./bedrock-server.zip -d "$SERVER_DIR"
+
+  sudo mv "$TEMP_DIR/server.properties" "$SERVER_DIR/server.properties"
+  sudo mv "$TEMP_DIR/allowlist.json" "$SERVER_DIR/allowlist.json"
+  sudo mv "$TEMP_DIR/permissions.json" "$SERVER_DIR/permissions.json"
+
+  sudo chown -R mcserver: /home/mcserver
+
+  sudo systemctl start mcbedrock
+
+  echo "$CURRENT_VERSION" > "$VERSION_FILE"
+
+  echo "Minecraft server updated to version $CURRENT_VERSION."
+
+else
+  echo "Minecraft server is already up to date (version $CURRENT_VERSION)."
+fi
